requirements.txt
streamlit
groq
python-dotenv
langchain
langchain-groq
faiss-cpu
sentence-transformers
langchain-community